+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Warning to `pcs resource|stonith update` commands about not using agent
   self-validation feature when the resource is already misconfigured
   ([rhbz#2151524])
+- Add lib command `cluster_property.set_properties` to API v2
 
 ### Fixed
 - Graceful stopping pcsd service using `systemctl stop pcsd` command